# internal/controller/http/v1/apiserver/apierror/errors.go
package apierror

import (
	"encoding/json"
	"io"
)

// ErrorResponse represents an API error response.
//
// @name ErrorResponse
type ErrorResponse struct {
	Ok      bool   `json:"ok" default:"false"`
	Message string `json:"message"`
}

func NewErrorResponse(message string) ErrorResponse {
	return ErrorResponse{
		Message: message,
	}
}

func (e ErrorResponse) Error() string {
	return e.Message
}

func WriteErrorMessage(writer io.Writer, message string) (int, error) {
	buf, err := json.Marshal(ErrorResponse{
		Message: message,
	})

	if err != nil {
		return writer.Write([]byte(message))
	}

	return writer.Write(buf)
}